# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Focus",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "https://dev.api.focus.grupomobius.com"
        },
        {
            "url": "https://api.focus.grupomobius.com"
        }
    ],
    "paths": {
...
        "/budget-pending-task/{id}": {
            "put": {
                "tags": [
                    "budget-pending-task"
                ],
                "summary": "Updated budget pending task",
                "operationId": "updateBudgetPendingTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated budget pending task object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BudgetPendingTask"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BudgetPendingTask"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Operation not found"
                    }
                }
            }
        },
...
        "/colors/{id}": {
            "put": {
                "tags": [
                    "colors"
                ],
                "summary": "Updated color",
                "operationId": "updateColor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated color object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Color"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Color"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Color not found"
                    }
                }
            }
        },
...
        "/api/customers/getall": {
            "get": {
                "tags": [
                    "customers"
                ],
                "summary": "Get all customers",
                "operationId": "fd67d498ed291052603ddd8f042458e3",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "summary": "Get customer by ID",
                "operationId": "9dbdb0ec3a90ff608165986f2cf42fac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found."
                    }
                }
            }
        },
        "/customers/update/{id}": {
            "put": {
                "tags": [
                    "customers"
                ],
                "summary": "Updated customer",
                "operationId": "updateCustomer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated customer object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                }
            }
        },
...
        "/api/defleet-variables/getall": {
            "get": {
                "tags": [
                    "defleet-variables"
                ],
                "summary": "Get all defleet vairables",
                "operationId": "dde4121a85d0e67a150192fcb218d4b9",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefleetVariable"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/estimated-dates/{id}": {
            "put": {
                "tags": [
                    "estimated-dates"
                ],
                "summary": "Updated estimated date",
                "operationId": "updateEstimatedDate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id that to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated estimated date object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstimatedDate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstimatedDate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estimated date not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BudgetPendingTask": {
                "title": "Budget pending task model",
                "description": "Budget pending task model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "campa_id": {
                        "title": "Campa ID",
                        "description": "Campa ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "role_id": {
                        "title": "Role ID",
                        "description": "Role ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "pending_task_id": {
                        "title": "Pending Task ID",
                        "description": "Pending Task ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "state_budget_pending_task_id": {
                        "title": "State of Budget Pending Task ID",
                        "description": "State of Budget Pending Task ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "code_authorization": {
                        "title": "Code of Authorization",
                        "description": "Code of Authorization",
                        "type": "integer",
                        "format": "int32"
                    },
                    "observations": {
                        "title": "Observations",
                        "description": "Observations",
                        "type": "string"
                    },
                    "url": {
                        "title": "URL",
                        "description": "URL",
                        "type": "string",
                        "format": "url"
                    },
                    "approved_by": {
                        "title": "Approved by",
                        "description": "Approved by",
                        "type": "integer",
                        "format": "int64"
                    },
                    "declined_by": {
                        "title": "Declined by",
                        "description": "Declined by",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "comment": {
                        "title": "Comment",
                        "description": "Comment",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Color": {
                "title": "Color model",
                "description": "Color model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "title": "Deleted at",
                        "description": "When was deleted",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Customer": {
                "title": "Customer model",
                "description": "Customer model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "company_id": {
                        "title": "Company ID",
                        "description": "Company ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "province_id": {
                        "title": "Province ID",
                        "description": "Province ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string"
                    },
                    "cif": {
                        "title": "Cif",
                        "description": "Cif",
                        "type": "string"
                    },
                    "phone": {
                        "title": "Phone",
                        "description": "Phone",
                        "type": "integer",
                        "format": "int32"
                    },
                    "address": {
                        "title": "Address",
                        "description": "Address",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "DefleetVariable": {
                "title": "Defleet variable model",
                "description": "Defleet variable model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "company_id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "kms": {
                        "title": "Kms",
                        "description": "Kms",
                        "type": "integer",
                        "format": "int32"
                    },
                    "years": {
                        "title": "Years",
                        "description": "Years",
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "EstimatedDate": {
                "title": "Estimated Date model",
                "description": "Estimated Date model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "pending_task_id": {
                        "title": "Pending Task ID",
                        "description": "Pending Task ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "reason": {
                        "title": "Reason",
                        "description": "Reason",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "When was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "When was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}